# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CrowdRelay API
  version: 0.1.0
  description: Open-source fan acquisition, referral, rewards, event conversion and admission API.
servers:
- url: https://api.example.com/v1
  description: Replace with your production API origin.
- url: http://localhost:8080/v1
  description: Local development.
paths:
...
  /fans/confirm:
    post:
      operationId: confirmFan
      summary: Confirm inbox ownership or recover an active fan session
      description: The server derives idempotency from the action token, so clicking, scanning or pasting the same link safely replays the successful session exchange.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FanActionRequest'
      responses:
        '200':
          description: Fan activated and a private session established.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              $ref: '#/components/headers/FanSessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FanConfirmationResponse'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '413':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
  /fans/unsubscribe:
    post:
      operationId: unsubscribeFan
      summary: Revoke marketing consent and active fan sessions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FanActionRequest'
      responses:
        '200':
          description: Fan unsubscribed and the browser session removed.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              $ref: '#/components/headers/FanSessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FanUnsubscribeResponse'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '413':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
...
  /events/{slug}/interest:
    post:
      operationId: registerEventInterest
      summary: Register an authenticated fan interest in an event
      security:
      - fanSession: []
      parameters:
      - *id003
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventInterestRequest'
      responses:
        '200':
          description: Existing interest or idempotent replay.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventInterestResult'
        '201':
          description: A new event interest was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventInterestResult'
        '400': *id004
        '401': *id004
        '404': *id004
        '409': *id004
        '422': *id004
        '503': *id004
...
  /admin/admission/passes:
    post:
      operationId: issueAdmissionPass
      summary: Issue one pass from a limited event pool
      security:
      - adminBearer: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueAdmissionPassRequest'
      responses:
        '201':
          description: Pass issued and winner delivery event recorded.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionPassIssued'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '401':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '413':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
  /admin/admission/passes/{public_reference}/revoke:
    post:
      operationId: revokeAdmissionPass
      summary: Revoke an unused admission pass
      security:
      - adminBearer: []
      parameters:
      - in: path
        name: public_reference
        required: true
        schema:
          type: string
          minLength: 8
          maxLength: 64
      - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Pass revoked and any winner session invalidated.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionPass'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '401':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
  /passes/claim:
    post:
      operationId: claimAdmissionPass
      summary: Exchange a one-time winner token for a private pass session
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimAdmissionPassRequest'
      responses:
        '200':
          description: Pass claimed and private session established.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
            Set-Cookie:
              $ref: '#/components/headers/PassSessionCookie'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionPass'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '413':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
...
  /staff/admission/redeem:
    post:
      operationId: redeemAdmissionPass
      summary: Atomically verify and redeem an admission pass
      security:
      - staffBearer: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedeemAdmissionPassRequest'
      responses:
        '200':
          description: Gate check completed. Repeated scans return the existing redemption state.
          headers:
            Cache-Control:
              $ref: '#/components/headers/PrivateNoStore'
            X-Request-ID:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdmissionRedemptionResult'
        '400':
          $ref: '#/components/responses/PrivateProblem'
        '401':
          $ref: '#/components/responses/PrivateProblem'
        '404':
          $ref: '#/components/responses/PrivateProblem'
        '409':
          $ref: '#/components/responses/PrivateProblem'
        '413':
          $ref: '#/components/responses/PrivateProblem'
        '422':
          $ref: '#/components/responses/PrivateProblem'
        '503':
          $ref: '#/components/responses/PrivateProblem'
        '500':
          $ref: '#/components/responses/PrivateProblem'
...
components:
  headers:
    RequestId:
      description: Correlation identifier generated by the API.
      schema:
        type: string
...
    FanSessionCookie:
      description: Opaque private fan session established after signup.
      required: true
      schema:
        type: string
        example: crowdrelay_fan=opaque; Max-Age=7776000; Path=/; Secure; HttpOnly; SameSite=Lax
    PrivateNoStore:
      description: Prevents storage of redirect, personal, or error responses.
      required: true
      schema:
        type: string
        const: private, no-store
...
    PassSessionCookie:
      description: Opaque winner session established after a one-time pass claim.
      required: true
      schema:
        type: string
        example: crowdrelay_pass_session=opaque; Max-Age=1209600; Path=/; Secure; HttpOnly; SameSite=Lax
...
  securitySchemes:
    synesthesiaRunBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-run-token
      description: Ephemeral bearer token returned by the Synesthesia run-start endpoint.
    fanSession:
      type: apiKey
      in: cookie
      name: crowdrelay_fan
    commerceBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
    ticketOrderBearer:
      type: http
      scheme: bearer
      bearerFormat: checkout-token
    passSession:
      type: apiKey
      in: cookie
      name: crowdrelay_pass_session
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
    staffBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
...
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: Caller-generated key scoped to this write operation. The same key and payload can be retried
        safely.
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
...
    PrivateProblem:
      description: RFC 9457 Problem Details response containing private request context.
      headers:
        Cache-Control:
          $ref: '#/components/headers/PrivateNoStore'
        X-Request-ID:
          $ref: '#/components/headers/RequestId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string
...
    EventInterestRequest:
      type: object
      additionalProperties: false
      properties:
        campaign_id:
          type:
          - string
          - 'null'
          format: uuid
        source:
          type: string
          minLength: 1
          maxLength: 128
          default: event_page
    EventInterestResult:
      type: object
      additionalProperties: false
      required:
      - event_id
      - fan_id
      - created
      - reminder_count
      properties:
        event_id:
          type: string
          format: uuid
        fan_id:
          type: string
          format: uuid
        created:
          type: boolean
        reminder_count:
          type: integer
          minimum: 0
...
    FanActionRequest:
      type: object
      additionalProperties: false
      required:
      - token
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 4096
          description: Raw 64-character action token or a confirmation URL containing token in its query string or fragment.
    FanConfirmationResponse:
      type: object
      additionalProperties: false
      required:
      - fan_id
      - status
      - referral_url
      - fan_session_token
      properties:
        fan_id:
          type: string
          format: uuid
        status:
          type: string
          const: active
        referral_url:
          type: string
          format: uri
        fan_session_token:
          type: string
          minLength: 24
          maxLength: 4096
          description: Opaque native-client session credential. Browser clients should continue using the HttpOnly cookie.
    FanUnsubscribeResponse:
      type: object
      additionalProperties: false
      required:
      - fan_id
      - status
      properties:
        fan_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - unsubscribed
          - suppressed
    AdmissionPassStatus:
      type: string
      enum:
      - issued
      - claimed
      - redeemed
      - revoked
      - expired
    AdmissionPass:
      type: object
      additionalProperties: false
      required:
      - pass_id
      - session_id
      - event_id
      - event_slug
      - event_title
      - venue
      - starts_at
      - holder_name
      - holder_email_masked
      - public_reference
      - status
      - session_expires_at
      - redeemed_at
      properties:
        pass_id:
          type: string
          format: uuid
        session_id:
          type:
          - string
          - 'null'
          format: uuid
        event_id:
          type: string
          format: uuid
        event_slug:
          type: string
        event_title:
          type: string
        venue:
          type:
          - string
          - 'null'
        starts_at:
          type: string
          format: date-time
        holder_name:
          type:
          - string
          - 'null'
        holder_email_masked:
          type: string
        public_reference:
          type: string
          minLength: 8
          maxLength: 64
          pattern: ^[A-Za-z0-9-]+$
        status:
          $ref: '#/components/schemas/AdmissionPassStatus'
        session_expires_at:
          type: string
          format: date-time
        redeemed_at:
          type:
          - string
          - 'null'
          format: date-time
    IssueAdmissionPassRequest:
      type: object
      additionalProperties: false
      required:
      - event_slug
      - pool_slug
      - fan_email
      properties:
        event_slug:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-z0-9][a-z0-9_-]*$
        pool_slug:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-z0-9][a-z0-9_-]*$
        fan_email:
          type: string
          format: email
        claim_expires_hours:
          type: integer
          minimum: 1
          maximum: 720
          default: 72
    AdmissionPassIssued:
      type: object
      additionalProperties: false
      required:
      - pass_id
      - event_id
      - fan_id
      - public_reference
      - claim_token
      - claim_expires_at
      - created
      properties:
        pass_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        fan_id:
          type: string
          format: uuid
        public_reference:
          type: string
        claim_token:
          type: string
          minLength: 64
          maxLength: 64
        claim_expires_at:
          type: string
          format: date-time
        created:
          type: boolean
    ClaimAdmissionPassRequest:
      type: object
      additionalProperties: false
      required:
      - token
      properties:
        token:
          type: string
          minLength: 64
          maxLength: 64
          pattern: ^[A-Fa-f0-9]{64}$
...
    RedeemAdmissionPassRequest:
      type: object
      additionalProperties: false
      required:
      - event_slug
      minProperties: 2
      maxProperties: 2
      properties:
        event_slug:
          type: string
          minLength: 1
          maxLength: 128
          pattern: ^[a-z0-9][a-z0-9_-]*$
        qr_token:
          type: string
          minLength: 1
          maxLength: 8192
        public_reference:
          type: string
          minLength: 8
          maxLength: 64
          pattern: ^[A-Za-z0-9-]+$
    AdmissionRedemptionResult:
      type: object
      additionalProperties: false
      required:
      - pass_id
      - event_id
      - public_reference
      - holder_name
      - holder_email_masked
      - status
      - redeemed_at
      properties:
        pass_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        public_reference:
          type: string
        holder_name:
          type:
          - string
          - 'null'
        holder_email_masked:
          type: string
        status:
          type: string
          enum:
          - redeemed
          - already_redeemed
          - revoked
          - expired
          - not_claimed
        redeemed_at:
          type:
          - string
          - 'null'
          format: date-time